Split store.ts into per-domain slices
Break the 1598-line src/store.ts into a src/store/ directory using the
Zustand slices pattern: app, workspace, repos, git, composer, drafts, ui,
account, appearance, and providers, with a types.ts composing AppState and
an index.ts barrel that preserves the public export surface (useAppStore,
AppState, DraftAgent, EMPTY_AGENTS, ChatItem, AgentUsage). All cross-slice
calls keep working through the unified get()/set(); behavior is unchanged.

Also relocate four misnamed store.*.test.ts files to sit next to the code
they actually exercise (helpers.ts and storage/preferences.ts).
